[Step] Runtime type checking: part 1
This is what it looks like to add additional runtime type checking. But
the generated code is still generating a more specific type than the
native side actually guarantees -- which implies that, for now, we
should make the TS type more generic

In addition, now we will use exported type in app.